requirements.txt
numpy
scipy
matplotlib
pandas
libroadrunner>=2.2.0
jupyter